Replace Concourse pipeline to run monitor with GitHub Actions workflow
We want to use GitHub Actions everywhere, instead of having some of the
infra running on a Concourse installation.

This commit replaces the Concourse CI pipeline with a GitHub Actions
workflow.

Note that although the schedules are configured to be similar, they
won't be exactly the same, as concourses' time resource behaves
differently to cron. Also, the GitHub Actions scheduled event doesn't
allow configuration of timezone, so the midday-ness of the new scheduled
messages will be subject to daylight savings.