Bump Ruby version to 2.7.5 and Bundler to 2.3.7
This is to work around an [AWS SAM CLI issue #76](https://github.com/aws/aws-lambda-builders/issues/76) whereby certain versions of Ruby simply don't seem to work. 2.7.4 was affected by this whereas 2.7.5 builds without issue. Given that 2.7.5 is the current stable 2.7 release it seemed sensible to upgrade to it.

Also bumped Bundler to a more modern version.